features/lib - Move the configuration properties to the WireMock module.
