+++ api-test-proxy/src/main/kotlin/io/fries/api/test/ApiTestExtension.kt
@@ -8,11 +8,10 @@ import org.junit.jupiter.api.extension.ExtensionContext
 import org.springframework.test.context.junit.jupiter.SpringExtension
 
 data class ApiTestContext(
-    private val recordPath: String,
     private val className: String,
     private val testName: String
 ) {
-    override fun toString(): String = "$recordPath/$className/$testName"
+    override fun toString(): String = "$className/$testName"
 }
 
 class ApiTestExtension : BeforeEachCallback, AfterEachCallback {
@@ -21,12 +20,10 @@ class ApiTestExtension : BeforeEachCallback, AfterEachCallback {
 
     override fun beforeEach(extensionContext: ExtensionContext) {
         val applicationContext = SpringExtension.getApplicationContext(extensionContext)
-        val apiTestProperties = applicationContext.getBean(ApiTestProperties::class.java)
         val proxyServersFactory = applicationContext.getBean(ProxyServersFactory::class.java)
 
         proxyServers = proxyServersFactory.create(
             ApiTestContext(
-                recordPath = apiTestProperties.recordPath,
                 className = extensionContext.requiredTestClass.canonicalName,
                 testName = extensionContext.requiredTestMethod.name
             )
